refactor: remove cache-layer, add comment
