+++ gateway/src/api/auth_layer.rs
@@ -21,14 +21,16 @@ use tower::{Layer, Service};
 use tracing::{error, trace, Span};
 use tracing_opentelemetry::OpenTelemetrySpanExt;
 
-use super::cache_layer::CacheManagement;
+use super::cache::CacheManagement;
 
 static PROXY_CLIENT: Lazy<ReverseProxy<HttpConnector<GaiResolver>>> =
     Lazy::new(|| ReverseProxy::new(Client::new()));
 
 /// The idea of this layer is to do two things:
 /// 1. Forward all user related routes (`/login`, `/logout`, `/users/*`, etc) to our auth service
-/// 2. Upgrade all Authorization Bearer keys and session cookies to JWT tokens for internal communication inside and below gateway
+/// 2. Upgrade all Authorization Bearer keys and session cookies to JWT tokens for internal
+/// communication inside and below gateway, fetching the JWT token from a ttl-cache if it isn't expired,
+/// and inserting it in the cache if it isn't there.
 #[derive(Clone)]
 pub struct ShuttleAuthLayer {
     auth_uri: Uri,
