+++ gateway/src/api/mod.rs
@@ -1,4 +1,4 @@
 mod auth_layer;
-mod cache_layer;
+mod cache;
 
 pub mod latest;
